started on plot submodule
