+++ python/src/scipp/plot/__init__.py
@@ -1,0 +1,9 @@
+# SPDX-License-Identifier: GPL-3.0-or-later
+# Copyright (c) 2019 Scipp contributors (https://github.com/scipp)
+# @file
+# @author Neil Vaytet
+
+# flake8: noqa
+
+from .plot import plot
+from . import config
