+++ python/src/scipp/plot/config.py
@@ -1,0 +1,18 @@
+# SPDX-License-Identifier: GPL-3.0-or-later
+# Copyright (c) 2019 Scipp contributors (https://github.com/scipp)
+# @file
+# @author Neil Vaytet
+
+
+# The plotting backend
+backend = "plotly"
+
+# The colorbar properties
+cb = {"name": "viridis", "log": False, "min": None, "max": None,
+      "min_var": None, "max_var": None}
+
+# The default image height (in pixels)
+height = 600
+
+# The default image width (in pixels)
+width = 950
